Better error message admitted by core.AndRight
This should be reverted or else all other tactics in this package should
have a similar modification. There's a good argument to be made that this
error should be wrapped in the tactic layer instead of in the core, so
I'm leaving this as a change to a sginle prop rule and we can decide at code
review time where these match errors should be caught and wrapped.