Added refinement cases to substitution helper
@todo review this commit
This commit was drafted in order to fix a bug where assignb would fall back
to the equational version instead of performing the appropriate substitution
because the substitution helper failed. The actual error was masked because
assignb is an Either tactic.

To uncover the error I replaced assignb with (TactixLibrary.useAt("[:=] assign")('R))
in RefinementCalculusTests."paper example 3"